experiment with omitinsert property in field tags
When inserting new data to the table with autoincrement, it is not
desired to insert the ID too. On the other hand, when loading
the result from SELECT to the struct, ID should be loaded. This
means that this:
	ID int64 `db:"-"`
is not enough. This new property:
	ID int64 `db:",omitinsert"
solves that kind of problem.