dali: add support for prepared statements
Prepare method was addod to both DB and Tx types. Unlike the
database/sql Prepare method, this one also accepts arguments.
It is because some placeholders are not parameters (e.g. ?ident,
?raw), thus it is desirable to provide the way to use them.

Besides these placeholders, only ? placeholder is left for the
parameter binding. It seems difficult to print for example
?values placeholder without knowing what type is going to
be bound. Nevertheless, I am in favour of keeping it simple
anyway.